Verplaats zandloper.js en plaatjes passen op klein scherm
