+++ src/assets/vars.scss
@@ -17,16 +17,14 @@ $kleur-zwart: #000000;
   src: './fonts/MinionVariableConcept-Roman.otf';
 }
 
+
 $font-header: Minion Variable Concept, Arial, Helvetica, sans-serif;
 $font-text: Minion Variable Concept, Arial, Helvetica, sans-serif;
-
-
 
 $primary: $kleur-donkerrood;
 $secondary: $kleur-rood;
 $font-color-primary: $kleur-goud0;
 $font-color-secondary: #FFFFFF;
-
 
 $theme-colors: (
   "primary": $primary,
