+++ src/pages/home/home.html
@@ -22,21 +22,24 @@
         <img id="zandloper_heel" class="heel" src="${require('../../assets/images/ZandloperHeel.png')}"/>
     </div>
 
-    <a id="home"></a>
+
     <nav class="site-header">
-        <a class="logo-titel d-inline-block" href="/lustrum12">
+        <a class="logo-titel d-inline-block" href="">
             <img src="${require('../../assets/images/themaheader.png')}"
                     alt="Logo menu" />
         </a>
         <div class="container">
             <div class="menu d-flex flex-column flex-md-row justify-content-between align-items-center">
-                <a class="d-none d-md-inline-block" href="#home">Home</a>
+                <a class="d-none d-md-inline-block" href="#top">Top</a>
                 <!-- <a class="d-none d-md-inline-block" href="opening">Lustrum Opening</a> -->
-                <a class="d-none d-md-inline-block" href="lustrumweek">Lustrum Week</a>
-                <a class="d-none d-md-inline-block" href="thema">Thema</a>
+                <a class="d-none d-md-inline-block" href="#activiteiten">Activiteiten</a>
+                <a class="d-none d-md-inline-block lw-font-text" href="lustrumweek">Lustrum Week</a>
+                <a class="d-none d-md-inline-block" href="#instagram">Instagram</a>
             </div>
         </div>
     </nav>
+
+    <a class="anchor glow" id="top"></a>
 
     <div class="thema thema-container">
         <img class="thema-plaetje" src="${require('../../assets/images/thematitel.png')}" alt="Lustrum Instagram" >
@@ -46,6 +49,36 @@
         <img class="boven" src="${require('../../assets/images/DividerBoven.png')}">
     </div>
 
+
+    <div class="overzicht aankomend" >
+        <a class="anchor" id="activiteiten"></a>
+        <h2 class="titel">Aankomende Activiteiten</h2>
+        <div class="activiteit">
+            <img class="activiteit-plaetje" src="${require('../../assets/images/home-Lustrumopening.png')}" alt="Lustrum Week">
+        </div>
+        <div class="activiteit">
+            <a href="lustrumweek">
+                <img class="activiteit-plaetje" src="${require('../../assets/images/home-Lustrumweek.png')}" alt="Lustrum Week">
+            </a>
+        </div>
+    </div>
+
+    <div class="divider">
+        <img class="tussen" src="${require('../../assets/images/DividerTussen.png')}">
+    </div>
+
+    <div class="overzicht geweest">
+        <h2 class="titel">Vorige Activiteiten</h2>
+        <div class="activiteit">
+            <a href="thema">
+                <img class="activiteit-plaetje" src="${require('../../assets/images/home-Themapresentatie.png')}" alt="Thema Presentatie">
+            </a>
+        </div>
+    </div>
+
+    <div class="divider">
+        <img class="tussen" src="${require('../../assets/images/DividerTussen.png')}">
+    </div>
 
     <div class="instagram">
         <a class="anchor" id="instagram"></a>
@@ -59,9 +92,7 @@
         </div>
     </div>
     
-    <!-- <div class="divider">
-        <img class="tussen" src="${require('../../assets/images/DividerTussen.png')}">
-    </div> -->
+
     
     <div class="divider">
         <img class="onder" src="${require('../../assets/images/DividerOnder.png')}">
