+++ src/pages/lustrumweek/lustrumweek.html
@@ -14,23 +14,22 @@
 </head>
 
 <body>
-    <a id="home"></a>
-
-
-
     <nav class="site-header">
-        <a class="logo-titel d-inline-block" href="/lustrum12">
+        <a class="logo-titel d-inline-block" href="../">
             <img src="${require('../../assets/images/themaheader.png')}"
                     alt="Logo menu" />
         </a>
         <div class="container">
             <div class="menu d-flex flex-column flex-md-row justify-content-between align-items-center headerbuttons">
-                <a class="d-none d-md-inline-block" href="#home">Home</a>
+                <a class="d-none d-md-inline-block" href="../">Home</a>
+                <a class="d-none d-md-inline-block" href="#top">Top</a>
                 <a class="d-none d-md-inline-block" href="#programma">Programma</a>
                 <!-- <a class="d-none d-md-inline-block" href="#poster">Poster</a> -->
             </div>
         </div>
     </nav>
+
+    <a class="anchor glow" id="top"></a>
 
     <div class="lw-thema">
         <h1 class="lw-titel">stressloos</h1>
