+++ src/pages/lustrumweek/lustrumweek.scss
@@ -1,30 +1,7 @@
 @import '../../assets/style.scss';
+@import './lustrumweek-vars.scss';
 
-$lw-blauw: #3B9596;
-$lw-oranje: #EC673E;
-$lw-donkerrood:   #551D10;      //#67250D;
-$lw-header-donkerrood:  #431413;    //#310c16;
 
-$lw-beige: #F2C292;
-
-$lw-beige-background: #f2c292e1;
-$lw-primary: $lw-blauw;
-$lw-secondary: $lw-oranje;
-
-@font-face {
-    font-family: Segoe Script;
-    src: '../../assets/fonts/segoescb.ttf';
-}
-
-@font-face {
-    font-family: 'Mrs Sheppards Regular';
-    font-style: italic;
-    font-weight: normal;
-    src: local('Mrs Sheppards Regular'), url('../../assets/fonts/MrsSheppards-Regular.woff') format('woff');
-}
-
-$lw-font-header: Mrs Sheppards Regular, Arial, Helvetica, sans-serif;
-$lw-font-text: Segoe Script, Arial, Helvetica, sans-serif;
 
 body {
     background-image: url('../../assets/images/lw-achtergrond.png');
@@ -94,7 +71,7 @@ nav {
         font-size: 24px;
         transition: ease-in-out color .15s;
         &:hover {
-            color: $lw-donkerrood;
+            color: $lw-blauw;
             text-decoration: none;
         }
     }
@@ -158,7 +135,8 @@ nav {
 
             }
             .lw-beschrijving {
-                color: $lw-blauw;
+                color: $lw-blauw; //grijs;
+                font-weight: bold;
                 text-shadow: -2px 1px 0 $lw-beige-background,
                             1px 2px 0 $lw-beige-background,
                             2px -1px 0 $lw-beige-background,
