+++ src/pages/thema/thema.html
@@ -22,15 +22,15 @@
         <img id="zandloper_heel" class="heel" src="${require('../../assets/images/ZandloperHeel.png')}"/>
     </div>
 
-    <a id="home"></a>
     <nav class="site-header">
-        <a class="logo-titel d-inline-block" href="/lustrum12">
+        <a class="logo-titel d-inline-block" href="../">
             <img src="${require('../../assets/images/themaheader.png')}"
                     alt="Logo menu" />
         </a>
         <div class="container">
             <div class="menu d-flex flex-column flex-md-row justify-content-between align-items-center">
-                <a class="d-none d-md-inline-block" href="#home">Home</a>
+                <a class="d-none d-md-inline-block" href="../">Home</a>
+                <a class="d-none d-md-inline-block" href="#top">Top</a>
                 <a class="d-none d-md-inline-block" href="#themafilm">Themafilm</a>
                 <a class="d-none d-md-inline-block" href="#game">Game</a>
                 <a class="d-none d-md-inline-block" href="#instagram">Instagram</a>
@@ -39,6 +39,7 @@
         </div>
     </nav>
 
+    <a id="top"></a>
     <div class="thema thema-container">
         <img class="thema-plaetje" src="${require('../../assets/images/thematitel.png')}" alt="Lustrum Instagram" >
     </div>
