+++ src/pages/thema/thema.scss
@@ -6,14 +6,14 @@ margin-bottom: 30px;
 }
 /* game */
 .game .game-container {
-display: flex;
-justify-content: center;
+    display: flex;
+    justify-content: center;
 
-.unity-container {
-    width: 960px;
-    height: 650px;
-    border:0px;
-}
+    .unity-container {
+        width: 960px;
+        height: 650px;
+        border:0px;
+    }
 }
 
 /* Commissie */
@@ -25,6 +25,7 @@ justify-content: center;
 
     .commissiefoto {
         width: 720px;
+        max-width: 100%;
         margin-bottom: 10px;
         margin-top: 10px;
     }
